first refactor
