+++ tinywl/compile_commands.json
@@ -7,6 +7,8 @@
       "-g",
       "-ggdb",
       "-O1",
+      "-I.",
+      "-Iinclude",
       "-g",
       "-Werror",
       "-I.",
@@ -14,7 +16,7 @@
       "-I/nix/store/xxh34bcvbzqp9r0z98r2qmja3lp3qf7r-wlroots-0.17.3/include",
       "-I/nix/store/zxblsvnpxaf4yslyifnwww1kdgi1dmyc-wayland-1.22.0-dev/include",
       "-I/nix/store/afp4ycmcgijcq6alln90v5sbryw09q81-libxkbcommon-1.7.0-dev/include",
-      "-I/nix/store/ykzlrlb2zvz4qsngr16bv2140f637rd2-pixman-0.43.4/include",
+      "-I/nix/store/idqvgrnmvxrrr4mix4djcp3y54pzma9n-pixman-0.43.4/include/pixman-1",
       "-o",
       "libtinywl.so",
       "tinywl.c"
