--- tinywl/meson.build
@@ -1,5 +1,0 @@
-executable(
-	'tinywl',
-	['tinywl.c', protocols_server_header['xdg-shell']],
-	dependencies: wlroots,
-)
